tests(storage): move storage tests to separate file
